move RedisHolder.kt to common
