+++ src/main/kotlin/org/radarbase/push/integration/garmin/backfill/GarminRequestGenerator.kt
@@ -5,7 +5,7 @@ import org.radarbase.gateway.Config
 import org.radarbase.push.integration.common.user.User
 import org.radarbase.push.integration.garmin.backfill.route.*
 import org.radarbase.push.integration.garmin.user.GarminUserRepository
-import org.radarbase.push.integration.garmin.util.RedisHolder
+import org.radarbase.push.integration.common.redis.RedisHolder
 import org.radarbase.push.integration.garmin.util.offset.*
 import org.slf4j.LoggerFactory
 import redis.clients.jedis.JedisPool
