+++ src/main/kotlin/org/radarbase/push/integration/garmin/util/RedisRemoteLockManager.kt
@@ -1,13 +1,14 @@
 package org.radarbase.push.integration.garmin.util
 
+import org.radarbase.push.integration.common.redis.RedisHolder
 import org.slf4j.LoggerFactory
 import redis.clients.jedis.params.SetParams
 import java.time.Duration
 import java.util.*
 
 class RedisRemoteLockManager(
-        private val redisHolder: RedisHolder,
-        private val keyPrefix: String
+    private val redisHolder: RedisHolder,
+    private val keyPrefix: String
 ) : RemoteLockManager {
     private val uuid: String = UUID.randomUUID().toString()
