+++ src/main/kotlin/org/radarbase/push/integration/garmin/util/offset/OffsetRedisPersistence.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectWriter
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import org.radarbase.push.integration.garmin.util.RedisHolder
+import org.radarbase.push.integration.common.redis.RedisHolder
 import org.slf4j.LoggerFactory
 import java.io.IOException
 import java.nio.file.Path
